+++ internal-packages/hub-api-client/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@quri/hub-api-client",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "build:ts": "tsc",
+    "build": "pnpm run build:ts"
+  },
+  "dependencies": {
+    "zod": "^3.24.2"
+  },
+  "devDependencies": {
+    "@quri/configs": "workspace:*",
+    "@types/node": "^22.13.1",
+    "typescript": "^5.7.3"
+  },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
+  }
+}
