+++ internal-packages/hub-api-client/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@quri/configs/tsconfig.base.json",
+  "compilerOptions": {
+    "tsBuildInfoFile": "dist/build.tsbuildinfo"
+  }
+}
